+++ server/src/comm/serial.rs
@@ -11,7 +11,7 @@ use std::sync::mpsc;
 use super::CommChannelTx;
 
 
-pub fn create_serial_comm_task(serial_id: usize) -> (CommChannelTx, thread::JoinHandle<()>) {
+pub fn create_serial_comm_task(serial_id: u32) -> (CommChannelTx, thread::JoinHandle<()>) {
     let env_var_str = format!("SERIAL_PORT_{}_PATH", serial_id);
 
     let mut serial_port = serial::open(
